extract common resolution utils from ivy utils
