+++ src/python/pants/backend/jvm/BUILD
@@ -48,6 +48,29 @@ python_library(
     '3rdparty/python/twitter/commons:twitter.common.collections',
     '3rdparty/python:six',
     ':ivy_utils_resources',
+    ':resolution_utils',
+    'src/python/pants/backend/jvm/subsystems:jar_dependency_management',
+    'src/python/pants/backend/jvm/targets:jvm',
+    'src/python/pants/base:build_environment',
+    'src/python/pants/base:deprecated',
+    'src/python/pants/base:generator',
+    'src/python/pants/base:revision',
+    'src/python/pants/build_graph',
+    'src/python/pants/ivy',
+    'src/python/pants/java:util',
+    'src/python/pants/java/jar',
+    'src/python/pants/util:dirutil',
+    'src/python/pants/util:fileutil',
+  ],
+)
+
+python_library(
+  name='resolution_utils',
+  sources=['resolution_utils.py'],
+  dependencies=[
+    '3rdparty/python/twitter/commons:twitter.common.collections',
+    '3rdparty/python:six',
+    ':ivy_utils_resources',
     'src/python/pants/backend/jvm/subsystems:jar_dependency_management',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:build_environment',
