+++ tests/python/pants_test/backend/jvm/tasks/test_ivy_utils.py
@@ -124,23 +124,6 @@ class IvyUtilsGenerateIvyTest(IvyUtilsTestBase):
     self.c = self.target('src/java/targets:c')
     self.e = self.target('src/java/targets:e')
 
-  def test_exclude_exported(self):
-    jars, excludes = IvyUtils.calculate_classpath([self.b])
-    for jar in jars:
-      self.assertEqual(jar.excludes, (Exclude(self.b_org, self.b_name),))
-    self.assertEqual(excludes, set())
-
-  def test_excludes_are_generated(self):
-    _, excludes = IvyUtils.calculate_classpath([self.e])
-    self.assertSetEqual(excludes, {Exclude(org='commons-lang', name='commons-lang')})
-
-  def test_classifiers(self):
-    jars, _ = IvyUtils.calculate_classpath([self.c])
-
-    jars.sort(key=lambda jar: jar.classifier)
-
-    self.assertEquals(['fleem', 'morx'], [jar.classifier for jar in jars])
-
   def test_module_ref_str_minus_classifier(self):
     module_ref = IvyModuleRef(org='org', name='name', rev='rev')
     self.assertEquals("IvyModuleRef(org:name:rev::jar)", str(module_ref))
@@ -171,42 +154,6 @@ class IvyUtilsGenerateIvyTest(IvyUtilsTestBase):
 
       override = self.find_single(doc, 'dependencies/override')
       self.assert_attributes(override, org='org2', module='name2', rev='rev2')
-
-  def test_resolve_conflict_missing_versions(self):
-    v1 = JarDependency('org.example', 'foo', None, force=False)
-    v2 = JarDependency('org.example', 'foo', '2', force=False)
-    self.assertIs(v2, IvyUtils._resolve_conflict(v1, v2))
-    self.assertIs(v2, IvyUtils._resolve_conflict(v2, v1))
-
-  def test_resove_conflict_no_conflicts(self):
-    v1 = JarDependency('org.example', 'foo', '1', force=False)
-    v1_force = JarDependency('org.example', 'foo', '1', force=True)
-    v2 = JarDependency('org.example', 'foo', '2', force=False)
-
-    # If neither version is forced, use the latest version.
-    self.assertIs(v2, IvyUtils._resolve_conflict(v1, v2))
-    self.assertIs(v2, IvyUtils._resolve_conflict(v2, v1))
-
-    # If an earlier version is forced, use the forced version.
-    self.assertIs(v1_force, IvyUtils._resolve_conflict(v1_force, v2))
-    self.assertIs(v1_force, IvyUtils._resolve_conflict(v2, v1_force))
-
-    # If the same version is forced, use the forced version.
-    self.assertIs(v1_force, IvyUtils._resolve_conflict(v1, v1_force))
-    self.assertIs(v1_force, IvyUtils._resolve_conflict(v1_force, v1))
-
-    # If the same force is in play in multiple locations, allow it.
-    self.assertIs(v1_force, IvyUtils._resolve_conflict(v1_force, v1_force))
-
-  def test_resolve_conflict_conflict(self):
-    v1_force = JarDependency('org.example', 'foo', '1', force=True)
-    v2_force = JarDependency('org.example', 'foo', '2', force=True)
-
-    with self.assertRaises(IvyUtils.IvyResolveConflictingDepsError):
-      IvyUtils._resolve_conflict(v1_force, v2_force)
-
-    with self.assertRaises(IvyUtils.IvyResolveConflictingDepsError):
-      IvyUtils._resolve_conflict(v2_force, v1_force)
 
   def test_get_resolved_jars_for_coordinates(self):
     ivy_info = self.parse_ivy_report('ivy_utils_resources/report_with_diamond.xml')
